Created an InputAdapter class for handling keyboard events.
Created a header for SdlGameView and created constructors for SdlGameView, SdlGameViewAdapter, and SdlInputAdapter.

These constructors take pointers to a GameController and various other classes, removing redundancy from the code (i.e. we no longer create multiple copies of the same classes that are used in the same contexts)

Added methods to GameController that pass method calls from InputAdapter to NetworkClient, which will send packets to the server and handle changes to the GameState

Currently implementing changes to FakeNetworkClient to test passing user input into a new GameState via server.