Created the classes Shield and Generator for future game development.  Shields and Generators will play a vital role for our gameplay premise.
Initialized a vector lists in GameState for shields and generators.  Created functions to add more instances of objects to these vector lists.

Created the necessary methods to pass Shield and Generator information from GameState to GameController, from GameController to SdlGameViewAdapter, and from SdlGameViewAdapter to SdlGameView.

These objects aren't drawn or rendered by SdlGameView as of yet, this is simply a framework so all the pieces work when we add these features in later.