Created a new class, Powerup.cpp and Powerup.h, which will be a container for powerup objects in GameState.
Powerup contains an x, y, and radius field for drawing on the game screen.  It also includes an int field, type, which will be used to determine which type of powerup it is.

GameState now contains a vector list of Powerup objects.

Methods have been created to add powerups to the list in GameState, and to pass the powerups list from GameState to GameController to GameViewAdapter and into GameView.

Powerups currently are drawn as different colored squares depending on the powerup type, this can be changed later.

Vehicle class needs updated with fields for handling powerups.

JSONEncodeDecode needs handles for passing powerups from serverside