Added new method to SdlGameView called drawHUD.
drawHUD takes the clients vehicle, determined by the isMe boolean in the Vehicle class, to create a basic heads up display for the player.

Currently the structure exists to begin drawing player's health, numerically with font, at the top of the screen.

CMakeLists now links GLUT libraries to the project.

README has been updated with instructions for an install to fix an error where certain GLUT libraries are set to NOTFOUND in CMakeLists, which can prevent compiling the project.