Created a more dynamic coloring effect for shields in SdlGameView.
Shields now slowly pulsate, creating a more interesting visual.

An arrow now points to the client vehicle, so the user knows which vehicle they are controlling.

This arrow uses the same pulsate effect as shields to draw more attention to that vehicle.

Removed old glew code that was commented out and no longer needed.  Note that we should remove the glew library calls in CMake too.